@@ -17,17 +17,19 @@ namespace Server
 	{
 		public void ConfigureServices(IServiceCollection services)
 		{
-			services.TryAddSingleton<ISampleService, SampleService>();
-			services.AddMvc(x => x.EnableEndpointRouting = false);
 			services.AddSoapCore();
+			services.TryAddSingleton<ISampleService, SampleService>();
+			services.AddMvc();
 		}
 
-		public void Configure(IApplicationBuilder app, IHostingEnvironment env, ILoggerFactory loggerFactory)
+		public void Configure(IApplicationBuilder app)
 		{
-			app.UseSoapEndpoint<ISampleService>("/Service.svc", new BasicHttpBinding(), SoapSerializer.DataContractSerializer);
-			app.UseSoapEndpoint<ISampleService>("/Service.asmx", new BasicHttpBinding(), SoapSerializer.XmlSerializer);
+			app.UseRouting();
 
-			app.UseMvc();
+			app.UseEndpoints(endpoints => {
+				endpoints.UseSoapEndpoint<ISampleService>("/Service.svc", new SoapEncoderOptions(), SoapSerializer.DataContractSerializer);
+				endpoints.UseSoapEndpoint<ISampleService>("/Service.asmx", new SoapEncoderOptions(), SoapSerializer.XmlSerializer);
+			});
 		}
 	}
 }
