@@ -9,7 +9,7 @@ namespace Client
 {
 	public class Program
 	{
-		public static void Main(string[] args)
+		public static void Main()
 		{
 			var binding = new BasicHttpBinding();
 			var endpoint = new EndpointAddress(new Uri(string.Format("http://{0}:5050/Service.svc", Environment.MachineName)));
