@@ -29,7 +29,9 @@ namespace SoapCore.Tests.MessageInspectors
 			switch (InspectorStyle)
 			{
 				case InspectorStyle.MessageInspector:
+#pragma warning disable CS0612 // Type or member is obsolete
 					services.AddSoapMessageInspector(new MessageInspectorMock());
+#pragma warning restore CS0612 // Type or member is obsolete
 					break;
 				case InspectorStyle.MessageInspector2:
 					services.AddSoapMessageInspector(new MessageInspector2Mock());
@@ -39,21 +41,20 @@ namespace SoapCore.Tests.MessageInspectors
 			services.AddMvc();
 		}
 
-#if ASPNET_21
+#if !NETCOREAPP3_0_OR_GREATER
 		public void Configure(IApplicationBuilder app, IHostingEnvironment env, ILoggerFactory loggerFactory)
 		{
-			app.UseSoapEndpoint<TestService>("/Service.svc", new BasicHttpBinding(), SoapSerializer.DataContractSerializer);
+			app.UseSoapEndpoint<TestService>("/Service.svc", new SoapEncoderOptions(), SoapSerializer.DataContractSerializer);
 			app.UseMvc();
 		}
-#endif
-#if ASPNET_30
+#else
 		public void Configure(IApplicationBuilder app, IWebHostEnvironment env, ILoggerFactory loggerFactory)
 		{
 			app.UseRouting();
 
 			app.UseEndpoints(x =>
 			{
-				x.UseSoapEndpoint<TestService>("/Service.svc", new BasicHttpBinding(), SoapSerializer.DataContractSerializer);
+				x.UseSoapEndpoint<TestService>("/Service.svc", new SoapEncoderOptions(), SoapSerializer.DataContractSerializer);
 			});
 		}
 #endif
