@@ -1,8 +1,4 @@
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using DeepEqual.Syntax;
 using Moq;
 using SoapCore.Tests.Serialization.Models.DataContract;
@@ -22,25 +18,52 @@ namespace SoapCore.Tests.Serialization
 
 		[Theory]
 		[InlineData(SoapSerializer.XmlSerializer)]
-		public void TestMessageHeadersModelWithoutBody(SoapSerializer serializer)
+		public void TestMessageHeadersModelWithBody(SoapSerializer serializer)
 		{
 			var service = _fixture.GetSampleServiceClient(serializer);
-			var model = new MessageHeadersModel
+			var model = new MessageHeadersModelWithBody
 			{
-				Prop1 = "test"
+				Prop1 = Guid.NewGuid().ToString(),
+				Prop2 = Guid.NewGuid().ToString(),
+				Prop3 = Guid.NewGuid().ToString(),
+				Prop4 = Guid.NewGuid().ToString(),
+				Prop5 = Guid.NewGuid().ToString(),
+				Prop6 = Guid.NewGuid().ToString(),
+				Prop7 = Guid.NewGuid().ToString(),
+				Prop8 = Guid.NewGuid().ToString(),
+				Body1 = Guid.NewGuid().ToString(),
+				Body2 = Guid.NewGuid().ToString()
 			};
 
-			_fixture.ServiceMock.Setup(x => x.Get(It.IsAny<MessageHeadersModel>())).Callback((MessageHeadersModel m) =>
+			_fixture.ServiceMock.Setup(x => x.GetWithBody(It.IsAny<MessageHeadersModelWithBody>())).Callback((MessageHeadersModelWithBody m) =>
 			{
 				m.ShouldDeepEqual(model);
-			}).Returns(new MessageHeadersModel
+			}).Returns(new MessageHeadersModelWithBody()
 			{
-				Prop1 = model.Prop1
+				Prop1 = model.Prop1,
+				Prop2 = model.Prop2,
+				Prop3 = model.Prop3,
+				Prop4 = model.Prop4,
+				Prop5 = model.Prop5,
+				Prop6 = model.Prop6,
+				Prop7 = model.Prop7,
+				Prop8 = model.Prop8,
+				Body1 = model.Body1,
+				Body2 = model.Body2
 			});
 
-			var result = service.Get(model);
+			var result = service.GetWithBody(model);
 
 			Assert.Equal(model.Prop1, result.Prop1);
+			Assert.Equal(model.Prop2, result.Prop2);
+			Assert.Equal(model.Prop3, result.Prop3);
+			Assert.Equal(model.Prop4, result.Prop4);
+			Assert.Equal(model.Prop5, result.Prop5);
+			Assert.Equal(model.Prop6, result.Prop6);
+			Assert.Equal(model.Prop7, result.Prop7);
+			Assert.Equal(model.Prop8, result.Prop8);
+			Assert.Equal(model.Body1, result.Body1);
+			Assert.Equal(model.Body2, result.Body2);
 		}
 
 		[Theory]
@@ -51,9 +74,15 @@ namespace SoapCore.Tests.Serialization
 			var model = new MessageHeadersModelWithBodyAndNamespace
 			{
 				Prop1 = Guid.NewGuid().ToString(),
+				Prop2 = Guid.NewGuid().ToString(),
+				Prop3 = Guid.NewGuid().ToString(),
+				Prop4 = Guid.NewGuid().ToString(),
+				Prop5 = Guid.NewGuid().ToString(),
+				Prop6 = Guid.NewGuid().ToString(),
+				Prop7 = Guid.NewGuid().ToString(),
+				Prop8 = Guid.NewGuid().ToString(),
 				Body1 = Guid.NewGuid().ToString(),
-				Body2 = Guid.NewGuid().ToString(),
-				Prop2 = Guid.NewGuid().ToString()
+				Body2 = Guid.NewGuid().ToString()
 			};
 
 			_fixture.ServiceMock.Setup(x => x.GetWithBodyAndNamespace(It.IsAny<MessageHeadersModelWithBodyAndNamespace>())).Callback((MessageHeadersModelWithBodyAndNamespace m) =>
@@ -62,49 +91,61 @@ namespace SoapCore.Tests.Serialization
 			}).Returns(new MessageHeadersModelWithBodyAndNamespace()
 			{
 				Prop1 = model.Prop1,
+				Prop2 = model.Prop2,
+				Prop3 = model.Prop3,
+				Prop4 = model.Prop4,
+				Prop5 = model.Prop5,
+				Prop6 = model.Prop6,
+				Prop7 = model.Prop7,
+				Prop8 = model.Prop8,
 				Body1 = model.Body1,
-				Body2 = model.Body2,
-				Prop2 = model.Prop2
+				Body2 = model.Body2
 			});
 
 			var result = service.GetWithBodyAndNamespace(model);
 
 			Assert.Equal(model.Prop1, result.Prop1);
 			Assert.Equal(model.Prop2, result.Prop2);
+			Assert.Equal(model.Prop3, result.Prop3);
+			Assert.Equal(model.Prop4, result.Prop4);
+			Assert.Equal(model.Prop5, result.Prop5);
+			Assert.Equal(model.Prop6, result.Prop6);
+			Assert.Equal(model.Prop7, result.Prop7);
+			Assert.Equal(model.Prop8, result.Prop8);
 			Assert.Equal(model.Body1, result.Body1);
 			Assert.Equal(model.Body2, result.Body2);
 		}
 
 		[Theory]
 		[InlineData(SoapSerializer.XmlSerializer)]
-		public void TestMessageHeadersModelWithBody(SoapSerializer serializer)
+		public void TestMessageHeadersModelWithNamespace(SoapSerializer serializer)
 		{
 			var service = _fixture.GetSampleServiceClient(serializer);
-			var model = new MessageHeadersModelWithBody
+			var model = new MessageHeadersModelWithNamespace
 			{
 				Prop1 = Guid.NewGuid().ToString(),
-				Body1 = Guid.NewGuid().ToString(),
-				Body2 = Guid.NewGuid().ToString(),
-				Prop2 = Guid.NewGuid().ToString()
+				Prop2 = Guid.NewGuid().ToString(),
+				Prop3 = Guid.NewGuid().ToString(),
+				Prop4 = Guid.NewGuid().ToString()
 			};
 
-			_fixture.ServiceMock.Setup(x => x.GetWithBody(It.IsAny<MessageHeadersModelWithBody>())).Callback((MessageHeadersModelWithBody m) =>
+			_fixture.ServiceMock.Setup(x => x.GetWithNamespace(It.IsAny<MessageHeadersModelWithNamespace>())).Callback((MessageHeadersModelWithNamespace m) =>
 			{
 				m.ShouldDeepEqual(model);
-			}).Returns(new MessageHeadersModelWithBody()
+			}).Returns(new MessageHeadersModelWithNamespace
 			{
 				Prop1 = model.Prop1,
-				Body1 = model.Body1,
-				Body2 = model.Body2,
-				Prop2 = model.Prop2
+				Prop2 = model.Prop2,
+				Prop3 = model.Prop3,
+				Prop4 = model.Prop4
 			});
 
-			var result = service.GetWithBody(model);
+			var result = service.GetWithNamespace(model);
 
 			Assert.Equal(model.Prop1, result.Prop1);
 			Assert.Equal(model.Prop2, result.Prop2);
-			Assert.Equal(model.Body1, result.Body1);
-			Assert.Equal(model.Body2, result.Body2);
+			Assert.Equal(model.Prop3, result.Prop3);
+			Assert.Equal(model.Prop4, result.Prop4);
 		}
 	}
 }
