@@ -22,12 +22,11 @@ namespace SoapCore.Tests
 			// is registered in app startup.
 
 			// Arrange
-			var logger = new NullLoggerFactory().CreateLogger<SoapEndpointMiddleware>();
+			var logger = new NullLoggerFactory().CreateLogger<SoapEndpointMiddleware<CustomMessage>>();
 
 			SoapOptions options = new SoapOptions()
 			{
 				Path = "/Service.svc", // this is the path registered in app startup
-				Binding = new CustomBinding(),
 				EncoderOptions = new[]
 				{
 					new SoapEncoderOptions
@@ -42,7 +41,7 @@ namespace SoapCore.Tests
 				SoapSerializer = SoapSerializer.DataContractSerializer
 			};
 
-			SoapEndpointMiddleware soapCore = new SoapEndpointMiddleware(logger, (innerContext) => Task.FromResult(TaskStatus.RanToCompletion), options);
+			SoapEndpointMiddleware<CustomMessage> soapCore = new SoapEndpointMiddleware<CustomMessage>(logger, (innerContext) => Task.FromResult(TaskStatus.RanToCompletion), options);
 
 			var context = new DefaultHttpContext();
 			context.Request.Path = new PathString("/DynamicPath/Service.svc");
@@ -66,12 +65,11 @@ namespace SoapCore.Tests
 			// is registered in app startup but a single path-part is not implemented.
 
 			// Arrange
-			var logger = new NullLoggerFactory().CreateLogger<SoapEndpointMiddleware>();
+			var logger = new NullLoggerFactory().CreateLogger<SoapEndpointMiddleware<CustomMessage>>();
 
 			SoapOptions options = new SoapOptions()
 			{
 				Path = "/v1/Service.svc", // this is the multi-part path registered in app startup
-				Binding = new CustomBinding(),
 				EncoderOptions = new[]
 				{
 					new SoapEncoderOptions
@@ -86,7 +84,7 @@ namespace SoapCore.Tests
 				SoapSerializer = SoapSerializer.DataContractSerializer
 			};
 
-			SoapEndpointMiddleware soapCore = new SoapEndpointMiddleware(logger, (innerContext) => Task.FromResult(TaskStatus.RanToCompletion), options);
+			SoapEndpointMiddleware<CustomMessage> soapCore = new SoapEndpointMiddleware<CustomMessage>(logger, (innerContext) => Task.FromResult(TaskStatus.RanToCompletion), options);
 
 			var context = new DefaultHttpContext();
 			context.Request.Path = new PathString("/DynamicPath/Service.svc");
@@ -109,12 +107,11 @@ namespace SoapCore.Tests
 			// when the service is not registered in app startup (opted-in).
 
 			// Arrange
-			var logger = new NullLoggerFactory().CreateLogger<SoapEndpointMiddleware>();
+			var logger = new NullLoggerFactory().CreateLogger<SoapEndpointMiddleware<CustomMessage>>();
 
 			SoapOptions options = new SoapOptions()
 			{
 				Path = "/v1/Service.svc", // this is the multi-part path registered in app startup
-				Binding = new CustomBinding(),
 				EncoderOptions = new[]
 				{
 					new SoapEncoderOptions
@@ -129,7 +126,7 @@ namespace SoapCore.Tests
 				SoapSerializer = SoapSerializer.DataContractSerializer
 			};
 
-			SoapEndpointMiddleware soapCore = new SoapEndpointMiddleware(logger, (innerContext) => Task.FromResult(TaskStatus.RanToCompletion), options);
+			SoapEndpointMiddleware<CustomMessage> soapCore = new SoapEndpointMiddleware<CustomMessage>(logger, (innerContext) => Task.FromResult(TaskStatus.RanToCompletion), options);
 
 			var context = new DefaultHttpContext();
 			context.Request.Path = new PathString("/v1/Service.svc");
