@@ -1,9 +1,5 @@
 using System;
 using System.Collections.Concurrent;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Xml.Serialization;
 
 namespace SoapCore
@@ -15,7 +11,7 @@ namespace SoapCore
 		public static XmlSerializer GetXmlSerializer(Type elementType, string parameterName, string parameterNs)
 		{
 			var key = $"{elementType}|{parameterName}|{parameterNs}";
-			return CachedSerializers.GetOrAdd(key, _ => new XmlSerializer(elementType, null, new Type[0], new XmlRootAttribute(parameterName), parameterNs));
+			return CachedSerializers.GetOrAdd(key, _ => new XmlSerializer(elementType, null, Array.Empty<Type>(), new XmlRootAttribute(parameterName), parameterNs));
 		}
 	}
 }
