@@ -1,5 +1,6 @@
 using System;
 using System.ServiceModel.Channels;
+using System.Xml;
 using SoapCore.Extensibility;
 
 namespace SoapCore
@@ -8,7 +9,9 @@ namespace SoapCore
 	/// The default implementation of the fault provider when an unexpected exception occurs. This can be replaced or
 	/// extended by registering your own IFaultExceptionTransformer in the service collection on startup.
 	/// </summary>
-	public class DefaultFaultExceptionTransformer : IFaultExceptionTransformer
+	/// <typeparam name="T_MESSAGE">The message type.</typeparam>
+	public class DefaultFaultExceptionTransformer<T_MESSAGE> : IFaultExceptionTransformer
+		where T_MESSAGE : CustomMessage, new()
 	{
 		private readonly ExceptionTransformer _exceptionTransformer;
 
@@ -22,7 +25,7 @@ namespace SoapCore
 			_exceptionTransformer = exceptionTransformer;
 		}
 
-		public Message ProvideFault(Exception exception, MessageVersion messageVersion)
+		public Message ProvideFault(Exception exception, MessageVersion messageVersion, Message requestMessage, XmlNamespaceManager xmlNamespaceManager)
 		{
 			var bodyWriter = _exceptionTransformer == null ?
 				new FaultBodyWriter(exception, messageVersion) :
@@ -30,7 +33,13 @@ namespace SoapCore
 
 			var soapCoreFaultMessage = Message.CreateMessage(messageVersion, null, bodyWriter);
 
-			return new CustomMessage(soapCoreFaultMessage);
+			T_MESSAGE customMessage = new T_MESSAGE
+			{
+				Message = soapCoreFaultMessage,
+				NamespaceManager = xmlNamespaceManager
+			};
+
+			return customMessage;
 		}
 	}
 }
