@@ -3,16 +3,28 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Runtime.CompilerServices;
 using System.ServiceModel.Channels;
-using System.Text;
+using System.Xml;
 using Microsoft.AspNetCore.Http;
 
 namespace SoapCore
 {
 	internal static class HeadersHelper
 	{
+		private static readonly char[] ContentTypeSeparators = new[] { ';' };
+
 		[MethodImpl(MethodImplOptions.AggressiveInlining)]
-		public static string GetSoapAction(HttpContext httpContext, Message requestMessage, System.Xml.XmlDictionaryReader reader)
+		public static string GetSoapAction(HttpContext httpContext, ref Message message)
 		{
+			XmlDictionaryReader reader = null;
+			if (!message.IsEmpty)
+			{
+				MessageBuffer mb = message.CreateBufferedCopy(int.MaxValue);
+				Message responseMsg = mb.CreateMessage();
+				message = mb.CreateMessage();
+
+				reader = responseMsg.GetReaderAtBodyContents();
+			}
+
 			var soapAction = httpContext.Request.Headers["SOAPAction"].FirstOrDefault();
 			if (soapAction == "\"\"")
 			{
@@ -21,7 +33,8 @@ namespace SoapCore
 
 			if (string.IsNullOrEmpty(soapAction))
 			{
-				foreach (var headerItem in httpContext.Request.Headers["Content-Type"])
+				var contentTypes = GetContentTypes(httpContext);
+				foreach (string headerItem in contentTypes)
 				{
 					// I want to avoid allocation as possible as I can(I hope to use Span<T> or Utf8String)
 					// soap1.2: action name is in Content-Type(like 'action="[action url]"') or body
@@ -87,14 +100,26 @@ namespace SoapCore
 
 				if (string.IsNullOrEmpty(soapAction))
 				{
+					if (!string.IsNullOrEmpty(message.Headers.Action))
+					{
+						soapAction = message.Headers.Action;
+					}
+
+					if (string.IsNullOrEmpty(soapAction))
+					{
+						var headerInfo = message.Headers.FirstOrDefault(h => h.Name.ToLowerInvariant() == "action");
+
+						if (headerInfo != null)
+						{
+							soapAction = message.Headers.GetHeader<string>(headerInfo.Name, headerInfo.Namespace);
+						}
+					}
+				}
+
+				if (string.IsNullOrEmpty(soapAction) && reader != null)
+				{
 					soapAction = reader.LocalName;
 				}
-			}
-
-			if (soapAction.Contains('/'))
-			{
-				// soapAction may be a path. Therefore must take the action from the path provided.
-				soapAction = soapAction.Split('/').Last();
 			}
 
 			if (!string.IsNullOrEmpty(soapAction))
@@ -104,6 +129,39 @@ namespace SoapCore
 			}
 
 			return soapAction;
+		}
+
+		public static string GetTrimmedClearedSoapAction(string inSoapAction)
+		{
+			string trimmedAction = GetTrimmedSoapAction(inSoapAction);
+
+			if (trimmedAction.EndsWith("Request"))
+			{
+				int endIndex = trimmedAction.LastIndexOf('R');
+				string clearedAction = trimmedAction.Substring(0, endIndex);
+				return clearedAction;
+			}
+
+			return trimmedAction;
+		}
+
+		public static string GetTrimmedSoapAction(string inSoapAction)
+		{
+			string soapAction = inSoapAction;
+			if (soapAction.Contains('/'))
+			{
+				// soapAction may be a path. Therefore must take the action from the path provided.
+				soapAction = soapAction.Split('/').Last();
+			}
+
+			return soapAction;
+		}
+
+		private static IEnumerable<string> GetContentTypes(HttpContext httpContext)
+		{
+			// in a single header entry is possible to find several content-types separated by ';'
+			return httpContext.Request.Headers["Content-Type"]
+				.SelectMany(c => c.Split(ContentTypeSeparators, StringSplitOptions.RemoveEmptyEntries));
 		}
 	}
 }
