@@ -7,12 +7,14 @@ using System.Collections.Generic;
 using System.Globalization;
 using System.IO;
 using System.IO.Pipelines;
+using System.Linq;
 using System.Net.Http.Headers;
 using System.ServiceModel;
 using System.ServiceModel.Channels;
 using System.Text;
 using System.Threading.Tasks;
 using System.Xml;
+using Microsoft.AspNetCore.Http;
 
 namespace SoapCore.MessageEncoder
 {
@@ -23,47 +25,63 @@ namespace SoapCore.MessageEncoder
 		private const string XmlMediaType = "application/xml";
 
 		private readonly Encoding _writeEncoding;
+		private readonly bool _overwriteResponseContentType;
 		private readonly bool _optimizeWriteForUtf8;
+		private readonly bool _omitXmlDeclaration;
+		private readonly bool _indentXml;
+		private readonly bool _checkXmlCharacters;
 
-		public SoapMessageEncoder(MessageVersion version, Encoding writeEncoding, XmlDictionaryReaderQuotas quotas)
+		public SoapMessageEncoder(MessageVersion version, Encoding writeEncoding, bool overwriteResponseContentType, XmlDictionaryReaderQuotas quotas, bool omitXmlDeclaration, bool indentXml, bool checkXmlCharacters, XmlNamespaceManager xmlNamespaceOverrides, string bindingName, string portName, int maxSoapHeaderSize = SoapMessageEncoderDefaults.MaxSoapHeaderSizeDefault)
 		{
-			if (writeEncoding == null)
-			{
-				throw new ArgumentNullException(nameof(writeEncoding));
-			}
-
-			SoapMessageEncoderDefaults.ValidateEncoding(writeEncoding);
+			_indentXml = indentXml;
+			_omitXmlDeclaration = omitXmlDeclaration;
+			_checkXmlCharacters = checkXmlCharacters;
+			BindingName = bindingName;
+			PortName = portName;
 
 			_writeEncoding = writeEncoding;
 			_optimizeWriteForUtf8 = IsUtf8Encoding(writeEncoding);
 
+			_overwriteResponseContentType = overwriteResponseContentType;
+
 			MessageVersion = version ?? throw new ArgumentNullException(nameof(version));
 
-			MessageVersion = version;
-
 			ReaderQuotas = new XmlDictionaryReaderQuotas();
-			quotas.CopyTo(ReaderQuotas);
+			(quotas ?? XmlDictionaryReaderQuotas.Max).CopyTo(ReaderQuotas);
+			MaxSoapHeaderSize = maxSoapHeaderSize;
 
 			MediaType = GetMediaType(version);
-			ContentType = GetContentType(MediaType, writeEncoding);
+			CharSet = SoapMessageEncoderDefaults.EncodingToCharSet(writeEncoding);
+			ContentType = GetContentType(MediaType, CharSet);
+
+			XmlNamespaceOverrides = xmlNamespaceOverrides;
 		}
+
+		public string BindingName { get; }
+		public string PortName { get; }
 
 		public string ContentType { get; }
 
 		public string MediaType { get; }
 
+		public string CharSet { get; }
+
 		public MessageVersion MessageVersion { get; }
 
 		public XmlDictionaryReaderQuotas ReaderQuotas { get; }
 
-		public bool IsContentTypeSupported(string contentType)
+		public int MaxSoapHeaderSize { get; }
+
+		public XmlNamespaceManager XmlNamespaceOverrides { get; }
+
+		public bool IsContentTypeSupported(string contentType, bool checkCharset)
 		{
 			if (contentType == null)
 			{
 				throw new ArgumentNullException(nameof(contentType));
 			}
 
-			if (IsContentTypeSupported(contentType, ContentType, MediaType))
+			if (IsContentTypeSupported(contentType, ContentType, MediaType, checkCharset))
 			{
 				return true;
 			}
@@ -76,22 +94,22 @@ namespace SoapCore.MessageEncoder
 				const string atomMediaType = "application/atom+xml";
 				const string htmlMediaType = "text/html";
 
-				if (IsContentTypeSupported(contentType, rss1MediaType, rss1MediaType))
+				if (IsContentTypeSupported(contentType, rss1MediaType, rss1MediaType, checkCharset))
 				{
 					return true;
 				}
 
-				if (IsContentTypeSupported(contentType, rss2MediaType, rss2MediaType))
+				if (IsContentTypeSupported(contentType, rss2MediaType, rss2MediaType, checkCharset))
 				{
 					return true;
 				}
 
-				if (IsContentTypeSupported(contentType, htmlMediaType, atomMediaType))
+				if (IsContentTypeSupported(contentType, htmlMediaType, atomMediaType, checkCharset))
 				{
 					return true;
 				}
 
-				if (IsContentTypeSupported(contentType, atomMediaType, atomMediaType))
+				if (IsContentTypeSupported(contentType, atomMediaType, atomMediaType, checkCharset))
 				{
 					return true;
 				}
@@ -107,7 +125,7 @@ namespace SoapCore.MessageEncoder
 				throw new ArgumentNullException(nameof(pipeReader));
 			}
 
-			var stream = new PipeStream(pipeReader, false);
+			using var stream = pipeReader.AsStream(true);
 			return await ReadMessageAsync(stream, maxSizeOfHeaders, contentType);
 		}
 
@@ -118,18 +136,44 @@ namespace SoapCore.MessageEncoder
 				throw new ArgumentNullException(nameof(stream));
 			}
 
-			XmlReader reader = XmlDictionaryReader.CreateTextReader(stream, ReaderQuotas);
+			XmlReader reader;
+
+			var readEncoding = SoapMessageEncoderDefaults.ContentTypeToEncoding(contentType);
+
+			if (readEncoding == null)
+			{
+				// Fallback to default or writeEncoding
+				readEncoding = _writeEncoding;
+			}
+
+			var supportXmlDictionaryReader = SoapMessageEncoderDefaults.TryValidateEncoding(readEncoding, out _);
+
+			if (supportXmlDictionaryReader)
+			{
+				reader = XmlDictionaryReader.CreateTextReader(stream, readEncoding, ReaderQuotas, dictionaryReader => { });
+			}
+			else
+			{
+				var streamReaderWithEncoding = new StreamReader(stream, readEncoding);
+				var xmlReaderSettings = new XmlReaderSettings() { IgnoreWhitespace = true, DtdProcessing = DtdProcessing.Prohibit, CloseInput = true };
+				reader = XmlReader.Create(streamReaderWithEncoding, xmlReaderSettings);
+			}
 
 			Message message = Message.CreateMessage(reader, maxSizeOfHeaders, MessageVersion);
 
 			return Task.FromResult(message);
 		}
 
-		public virtual async Task WriteMessageAsync(Message message, PipeWriter pipeWriter)
+		public virtual async Task WriteMessageAsync(Message message, HttpContext httpContext, PipeWriter pipeWriter)
 		{
 			if (message == null)
 			{
 				throw new ArgumentNullException(nameof(message));
+			}
+
+			if (httpContext == null)
+			{
+				throw new ArgumentNullException(nameof(httpContext));
 			}
 
 			if (pipeWriter == null)
@@ -139,26 +183,38 @@ namespace SoapCore.MessageEncoder
 
 			ThrowIfMismatchedMessageVersion(message);
 
-			using var bufferTextWriter = new BufferTextWriter(pipeWriter, _writeEncoding);
-			using var xmlTextWriter = new XmlTextWriter(bufferTextWriter);
-
-			var xmlWriter = XmlDictionaryWriter.CreateDictionaryWriter(xmlTextWriter);
-
-			if (_optimizeWriteForUtf8)
+			//Custom string writer with custom encoding support
+			using (var stringWriter = new CustomStringWriter(_writeEncoding))
 			{
-				message.WriteMessage(xmlWriter);
-			}
-			else
-			{
-				xmlWriter.WriteStartDocument();
-				message.WriteMessage(xmlWriter);
-				xmlWriter.WriteEndDocument();
-			}
+				using (var xmlTextWriter = XmlWriter.Create(stringWriter, new XmlWriterSettings
+				{
+					OmitXmlDeclaration = _optimizeWriteForUtf8 && _omitXmlDeclaration, //can only omit if utf-8
+					Indent = _indentXml,
+					Encoding = _writeEncoding,
+					CloseOutput = true,
+					CheckCharacters = _checkXmlCharacters
+				}))
+				{
+					using var xmlWriter = XmlDictionaryWriter.CreateDictionaryWriter(xmlTextWriter);
+					message.WriteMessage(xmlWriter);
+					xmlWriter.WriteEndDocument();
+					xmlWriter.Flush();
+				}
 
-			xmlWriter.Flush();
-			xmlWriter.Dispose();
+				var data = stringWriter.ToString();
+				var soapMessage = _writeEncoding.GetBytes(data);
 
-			await pipeWriter.FlushAsync();
+				//Set Content-length in Response
+				httpContext.Response.ContentLength = soapMessage.Length;
+
+				if (_overwriteResponseContentType)
+				{
+					httpContext.Response.ContentType = ContentType;
+				}
+
+				await pipeWriter.WriteAsync(soapMessage);
+				await pipeWriter.FlushAsync();
+			}
 		}
 
 		public virtual Task WriteMessageAsync(Message message, Stream stream)
@@ -175,22 +231,19 @@ namespace SoapCore.MessageEncoder
 
 			ThrowIfMismatchedMessageVersion(message);
 
-			XmlDictionaryWriter xmlWriter = XmlDictionaryWriter.CreateTextWriter(stream, _writeEncoding, false);
-
-			if (_optimizeWriteForUtf8)
+			using var xmlTextWriter = XmlWriter.Create(stream, new XmlWriterSettings
 			{
-				message.WriteMessage(xmlWriter);
-			}
-			else
-			{
-				xmlWriter.WriteStartDocument();
-				message.WriteMessage(xmlWriter);
-				xmlWriter.WriteEndDocument();
-			}
+				OmitXmlDeclaration = _optimizeWriteForUtf8 && _omitXmlDeclaration, //can only omit if utf-8,
+				Indent = _indentXml,
+				Encoding = _writeEncoding,
+				CloseOutput = false,
+				CheckCharacters = _checkXmlCharacters
+			});
 
+			using var xmlWriter = XmlDictionaryWriter.CreateDictionaryWriter(xmlTextWriter);
+			message.WriteMessage(xmlWriter);
+			xmlWriter.WriteEndDocument();
 			xmlWriter.Flush();
-
-			xmlWriter.Dispose();
 
 			return Task.CompletedTask;
 		}
@@ -219,93 +272,47 @@ namespace SoapCore.MessageEncoder
 			return mediaType;
 		}
 
-		internal static string GetContentType(string mediaType, Encoding encoding)
+		internal static string GetContentType(string mediaType, string charSet)
 		{
-			return string.Format(CultureInfo.InvariantCulture, "{0}; charset={1}", mediaType, SoapMessageEncoderDefaults.EncodingToCharSet(encoding));
+			return string.Format(CultureInfo.InvariantCulture, "{0}; charset={1}", mediaType, charSet);
 		}
 
-		internal bool IsContentTypeSupported(string contentType, string supportedContentType, string supportedMediaType)
+		internal bool IsContentTypeSupported(string contentType, string supportedContentType, string supportedMediaType, bool checkCharset)
 		{
 			if (supportedContentType == contentType)
 			{
 				return true;
 			}
 
-			if (contentType.Length > supportedContentType.Length &&
-				contentType.StartsWith(supportedContentType, StringComparison.Ordinal) &&
-				contentType[supportedContentType.Length] == ';')
+			MediaTypeHeaderValue parsedContentType;
+
+			try
 			{
-				return true;
+				parsedContentType = MediaTypeHeaderValue.Parse(contentType);
+			}
+			catch (FormatException)
+			{
+				//bad format
+				return false;
 			}
 
-			// now check case-insensitively
-			if (contentType.StartsWith(supportedContentType, StringComparison.OrdinalIgnoreCase))
+			if (parsedContentType.MediaType.Equals(MediaType, StringComparison.OrdinalIgnoreCase))
 			{
-				if (contentType.Length == supportedContentType.Length)
+				if (!checkCharset || string.IsNullOrWhiteSpace(parsedContentType.CharSet) || parsedContentType.CharSet.Equals(CharSet, StringComparison.OrdinalIgnoreCase))
 				{
 					return true;
-				}
-				else if (contentType.Length > supportedContentType.Length)
-				{
-					char ch = contentType[supportedContentType.Length];
-
-					// Linear Whitespace is allowed to appear between the end of one property and the semicolon.
-					// LWS = [CRLF]? (SP | HT)+
-					if (ch == ';')
-					{
-						return true;
-					}
-
-					// Consume the [CRLF]?
-					int i = supportedContentType.Length;
-					if (ch == '\r' && contentType.Length > supportedContentType.Length + 1 && contentType[i + 1] == '\n')
-					{
-						i += 2;
-						ch = contentType[i];
-					}
-
-					// Look for a ';' or nothing after (SP | HT)+
-					if (ch == ' ' || ch == '\t')
-					{
-						i++;
-						while (i < contentType.Length)
-						{
-							ch = contentType[i];
-							if (ch != ' ' && ch != '\t')
-							{
-								break;
-							}
-
-							++i;
-						}
-					}
-
-					if (ch == ';' || i == contentType.Length)
-					{
-						return true;
-					}
 				}
 			}
 
 			// sometimes we get a contentType that has parameters, but our encoders
 			// merely expose the base content-type, so we will check a stripped version
-			try
+			if (supportedMediaType.Length > 0 && !supportedMediaType.Equals(parsedContentType.MediaType, StringComparison.OrdinalIgnoreCase))
 			{
-				MediaTypeHeaderValue parsedContentType = MediaTypeHeaderValue.Parse(contentType);
-
-				if (supportedMediaType.Length > 0 && !supportedMediaType.Equals(parsedContentType.MediaType, StringComparison.OrdinalIgnoreCase))
-				{
-					return false;
-				}
-
-				if (!IsCharSetSupported(parsedContentType.CharSet))
-				{
-					return false;
-				}
+				return false;
 			}
-			catch (FormatException)
+
+			if (!IsCharSetSupported(parsedContentType.CharSet))
 			{
-				// bad content type, so we definitely don't support it!
 				return false;
 			}
 
@@ -314,7 +321,8 @@ namespace SoapCore.MessageEncoder
 
 		internal virtual bool IsCharSetSupported(string charset)
 		{
-			return false;
+			return CharSet?.Equals(charset, StringComparison.OrdinalIgnoreCase)
+				   ?? false;
 		}
 
 		private static bool IsUtf8Encoding(Encoding encoding)
