@@ -9,6 +9,7 @@ using System.ServiceModel;
 using System.ServiceModel.Channels;
 using System.Xml;
 using System.Xml.Serialization;
+using SoapCore.Meta;
 using SoapCore.ServiceModel;
 
 namespace SoapCore
@@ -22,8 +23,9 @@ namespace SoapCore
 		private readonly string _resultName;
 		private readonly object _result;
 		private readonly Dictionary<string, object> _outResults;
+		private readonly bool _skipResponseEnvelope;
 
-		public ServiceBodyWriter(SoapSerializer serializer, OperationDescription operation, object result, Dictionary<string, object> outResults) : base(isBuffered: true)
+		public ServiceBodyWriter(SoapSerializer serializer, OperationDescription operation, object result, Dictionary<string, object> outResults, bool skipResponseEnvelope = false) : base(isBuffered: true)
 		{
 			_serializer = serializer;
 			_operation = operation;
@@ -32,6 +34,7 @@ namespace SoapCore
 			_resultName = operation.ReturnName;
 			_result = result;
 			_outResults = outResults ?? new Dictionary<string, object>();
+			_skipResponseEnvelope = skipResponseEnvelope;
 		}
 
 		protected override void OnWriteBodyContents(XmlDictionaryWriter writer)
@@ -84,64 +87,14 @@ namespace SoapCore
 			// Do not wrap old-style single element response into additional xml element for xml serializer
 			var needResponseEnvelope = _result == null || (_outResults.Count > 0) || !_operation.IsMessageContractResponse;
 
+			if (_skipResponseEnvelope)
+			{
+				needResponseEnvelope = false;
+			}
+
 			if (needResponseEnvelope)
 			{
 				writer.WriteStartElement(_envelopeName, _serviceNamespace);
-			}
-
-			foreach (var outResult in _outResults)
-			{
-				string value = null;
-				if (outResult.Value is Guid)
-				{
-					value = outResult.Value.ToString();
-				}
-				else if (outResult.Value is bool)
-				{
-					value = outResult.Value.ToString().ToLower();
-				}
-				else if (outResult.Value is string)
-				{
-					value = System.Security.SecurityElement.Escape(outResult.Value.ToString());
-				}
-				else if (outResult.Value is Enum)
-				{
-					value = outResult.Value.ToString();
-				}
-				else if (outResult.Value == null)
-				{
-					value = null;
-				}
-				else
-				{
-					//for complex types
-					using (var ms = new MemoryStream())
-					using (var stream = new BufferedStream(ms))
-					{
-						// write element with name as outResult.Key and type information as outResultType
-						// i.e. <outResult.Key xsi:type="outResultType" ... />
-						var outResultType = outResult.Value.GetType();
-						var serializer = CachedXmlSerializer.GetXmlSerializer(outResultType, outResult.Key, _serviceNamespace);
-						lock (serializer)
-						{
-							serializer.Serialize(stream, outResult.Value);
-						}
-
-						//add outResultType. ugly, but working
-						stream.Position = 0;
-						XmlDocument xdoc = new XmlDocument();
-						xdoc.Load(stream);
-						var attr = xdoc.CreateAttribute("xsi", "type", "http://www.w3.org/2001/XMLSchema-instance");
-						attr.Value = outResultType.Name;
-						xdoc.DocumentElement.Attributes.Prepend(attr);
-						writer.WriteRaw(xdoc.DocumentElement.OuterXml);
-					}
-				}
-
-				if (value != null)
-				{
-					writer.WriteRaw(string.Format("<{0}>{1}</{0}>", outResult.Key, value));
-				}
 			}
 
 			if (_result != null)
@@ -152,11 +105,13 @@ namespace SoapCore
 				var xmlRootAttr = resultType.GetTypeInfo().GetCustomAttributes<XmlRootAttribute>().FirstOrDefault();
 				var messageContractAttribute = resultType.GetTypeInfo().GetCustomAttribute<MessageContractAttribute>();
 
-				var xmlName = _operation.ReturnElementName
+				var xmlName = _operation.ReturnElementName ?? messageContractAttribute?.WrapperName
 					?? (needResponseEnvelope
-					? _resultName
+					? (string.IsNullOrWhiteSpace(xmlRootAttr?.ElementName)
+						? _resultName
+						: xmlRootAttr.ElementName)
 					: (string.IsNullOrWhiteSpace(xmlRootAttr?.ElementName)
-					? resultType.Name
+					? BodyWriterExtensions.GetSerializedTypeName(resultType)
 					: xmlRootAttr.ElementName));
 
 				var xmlNs = _operation.ReturnNamespace ?? messageContractAttribute?.WrapperNamespace
@@ -164,16 +119,26 @@ namespace SoapCore
 					? _serviceNamespace
 					: xmlRootAttr.Namespace);
 
+				if (_operation.ReturnsChoice)
+				{
+					foreach (var ch in _operation.ReturnChoices)
+					{
+						if (ch.Type == resultType)
+						{
+							xmlName = ch.Name ?? xmlName;
+							xmlNs = ch.Namespace ?? xmlNs;
+							break;
+						}
+					}
+				}
+
 				var xmlArrayAttr = _operation.DispatchMethod.GetCustomAttribute<XmlArrayAttribute>();
 
 				if (xmlArrayAttr != null && resultType.IsArray)
 				{
 					var serializer = CachedXmlSerializer.GetXmlSerializer(resultType.GetElementType(), xmlName, xmlNs);
 
-					lock (serializer)
-					{
-						serializer.SerializeArray(writer, (object[])_result);
-					}
+					serializer.SerializeArray(writer, (object[])_result);
 				}
 				else
 				{
@@ -204,20 +169,17 @@ namespace SoapCore
 
 							var serializer = CachedXmlSerializer.GetXmlSerializer(memberType, memberName, memberNamespace);
 
-							lock (serializer)
+							if (memberValue is Stream)
 							{
-								if (memberValue is Stream)
-								{
-									writer.WriteStartElement(memberName, _serviceNamespace);
+								writer.WriteStartElement(memberName, _serviceNamespace);
 
-									WriteStream(writer, memberValue);
+								WriteStream(writer, memberValue);
 
-									writer.WriteEndElement();
-								}
-								else
-								{
-									serializer.Serialize(writer, memberValue);
-								}
+								writer.WriteEndElement();
+							}
+							else
+							{
+								serializer.Serialize(writer, memberValue);
 							}
 						}
 
@@ -230,20 +192,81 @@ namespace SoapCore
 					{
 						var serializer = CachedXmlSerializer.GetXmlSerializer(resultType, xmlName, xmlNs);
 
-						lock (serializer)
+						if (_result is Stream)
 						{
-							if (_result is Stream)
+							writer.WriteStartElement(_resultName, _serviceNamespace);
+							WriteStream(writer, _result);
+							writer.WriteEndElement();
+						}
+						else
+						{
+							//https://github.com/DigDes/SoapCore/issues/385
+							if (_operation.DispatchMethod.GetCustomAttribute<XmlSerializerFormatAttribute>()?.Style == OperationFormatStyle.Rpc)
 							{
-								writer.WriteStartElement(_resultName, _serviceNamespace);
-								WriteStream(writer, _result);
-								writer.WriteEndElement();
+								var importer = new SoapReflectionImporter(_serviceNamespace);
+								var typeMapping = importer.ImportTypeMapping(resultType);
+								var accessor = typeMapping.GetType().GetProperty("Accessor", BindingFlags.Instance | BindingFlags.Public | BindingFlags.NonPublic)?.GetValue(typeMapping);
+								accessor?.GetType().GetProperty("Name", BindingFlags.Instance | BindingFlags.Public | BindingFlags.NonPublic)?.SetValue(accessor, xmlName);
+								new XmlSerializer(typeMapping).Serialize(writer, _result);
 							}
 							else
 							{
+								//https://github.com/DigDes/SoapCore/issues/719
 								serializer.Serialize(writer, _result);
 							}
 						}
 					}
+				}
+			}
+
+			foreach (var outResult in _outResults)
+			{
+				string value = null;
+				if (outResult.Value is Guid)
+				{
+					value = outResult.Value.ToString();
+				}
+				else if (outResult.Value is bool)
+				{
+					value = outResult.Value.ToString().ToLower();
+				}
+				else if (outResult.Value is string)
+				{
+					value = System.Security.SecurityElement.Escape(outResult.Value.ToString());
+				}
+				else if (outResult.Value is Enum)
+				{
+					value = outResult.Value.ToString();
+				}
+				else if (outResult.Value == null)
+				{
+					value = null;
+				}
+				else
+				{
+					//for complex types
+					using (var stream = new MemoryStream())
+					{
+						// write element with name as outResult.Key and type information as outResultType
+						// i.e. <outResult.Key xsi:type="outResultType" ... />
+						var outResultType = outResult.Value.GetType();
+						var serializer = CachedXmlSerializer.GetXmlSerializer(outResultType, outResult.Key, _serviceNamespace);
+						serializer.Serialize(stream, outResult.Value);
+
+						//add outResultType. ugly, but working
+						stream.Position = 0;
+						XmlDocument xdoc = new XmlDocument();
+						xdoc.Load(stream);
+						var attr = xdoc.CreateAttribute("xsi", "type", Namespaces.XMLNS_XSI);
+						attr.Value = outResultType.Name;
+						xdoc.DocumentElement.Attributes.Prepend(attr);
+						writer.WriteRaw(xdoc.DocumentElement.OuterXml);
+					}
+				}
+
+				if (value != null)
+				{
+					writer.WriteRaw(string.Format("<{0}>{1}</{0}>", outResult.Key, value));
 				}
 			}
 
@@ -258,6 +281,31 @@ namespace SoapCore
 			Debug.Assert(_outResults != null, "Object should set empty out results");
 
 			writer.WriteStartElement(_envelopeName, _serviceNamespace);
+
+			if (_result != null)
+			{
+				if (_result is Stream)
+				{
+					writer.WriteStartElement(_resultName, _serviceNamespace);
+					WriteStream(writer, _result);
+					writer.WriteEndElement();
+				}
+				else
+				{
+					// When operation return type is `System.Object` the `DataContractSerializer` adds `i:type` attribute with the correct object type
+					Type resultType = _operation.ReturnType;
+					IEnumerable<Type> serviceKnownTypes = _operation
+						.GetServiceKnownTypesHierarchy()
+						.Select(x => x.Type);
+
+					// When `KnownTypeAttribute` is present the `DataContractSerializer` adds `i:type` attribute with the correct object type
+					DataContractSerializer serializer = resultType.TryGetBaseTypeWithKnownTypes(out Type resultBaseTypeWithKnownTypes)
+						? new DataContractSerializer(resultBaseTypeWithKnownTypes, _resultName, _serviceNamespace, serviceKnownTypes)
+						: new DataContractSerializer(resultType, _resultName, _serviceNamespace, serviceKnownTypes);
+
+					serializer.WriteObject(writer, _result);
+				}
+			}
 
 			foreach (var outResult in _outResults)
 			{
@@ -286,10 +334,16 @@ namespace SoapCore
 				else
 				{
 					//for complex types
-					using (var ms = new MemoryStream())
-					using (var stream = new BufferedStream(ms))
+					using (var stream = new MemoryStream())
 					{
-						new DataContractSerializer(outResult.Value.GetType()).WriteObject(ms, outResult.Value);
+						Type outResultType = outResult.Value.GetType();
+						IEnumerable<Type> serviceKnownTypes = _operation
+							.GetServiceKnownTypesHierarchy()
+							.Select(x => x.Type);
+
+						var serializer = new DataContractSerializer(outResultType, serviceKnownTypes);
+						serializer.WriteObject(stream, outResult.Value);
+
 						stream.Position = 0;
 						using (var reader = XmlReader.Create(stream))
 						{
@@ -302,21 +356,6 @@ namespace SoapCore
 				if (value != null)
 				{
 					writer.WriteRaw(string.Format("<{0}>{1}</{0}>", outResult.Key, value));
-				}
-			}
-
-			if (_result != null)
-			{
-				if (_result is Stream)
-				{
-					writer.WriteStartElement(_resultName, _serviceNamespace);
-					WriteStream(writer, _result);
-					writer.WriteEndElement();
-				}
-				else
-				{
-					var serializer = new DataContractSerializer(_result.GetType(), _resultName, _serviceNamespace);
-					serializer.WriteObject(writer, _result);
 				}
 			}
 
