@@ -11,6 +11,7 @@ namespace SoapCore.ServiceModel
 		public ServiceDescription(Type serviceType)
 		{
 			ServiceType = serviceType;
+			ServiceKnownTypes = serviceType.GetCustomAttributes<ServiceKnownTypeAttribute>(inherit: false);
 
 			var types = Enumerable.Empty<Type>().Concat(ServiceType.GetInterfaces());
 			types = types.Concat(new[] { ServiceType });
@@ -20,15 +21,34 @@ namespace SoapCore.ServiceModel
 			{
 				foreach (var serviceContract in contractType.GetTypeInfo().GetCustomAttributes<ServiceContractAttribute>())
 				{
-					contracts.Add(new ContractDescription(this, contractType, serviceContract));
+					var contractDescription = new ContractDescription(this, contractType, serviceContract);
+
+					contracts.Add(contractDescription);
+
+					if (GeneralContract is null)
+					{
+						GeneralContract = contractDescription;
+					}
+					else
+					{
+						if (GeneralContract.GetType().IsAssignableFrom(contractDescription.GetType()))
+						{
+							GeneralContract = contractDescription;
+						}
+					}
 				}
 			}
 
 			Contracts = contracts;
+
+			ServiceName = GeneralContract?.Name ?? serviceType.Name;
 		}
 
-		public Type ServiceType { get; private set; }
-		public IEnumerable<ContractDescription> Contracts { get; private set; }
+		public Type ServiceType { get; }
+		public string ServiceName { get; }
+		public ContractDescription GeneralContract { get; }
+		public IEnumerable<ServiceKnownTypeAttribute> ServiceKnownTypes { get; }
+		public IEnumerable<ContractDescription> Contracts { get; }
 		public IEnumerable<OperationDescription> Operations => Contracts.SelectMany(c => c.Operations);
 	}
 }
